rename wincam to wincamd for the constants
