+++ src/cameras/wincamd_constants.py
